Extract the SQL editor into QueryEditorPanel UserControl
Continues the MainWindow decomposition (UI design rule 7). The SQL editor —
the single biggest cluster of interaction logic left in MainWindow's
code-behind — moves into a focused QueryEditorPanel UserControl.

What moved (~670 lines of code-behind):
- The AvaloniaEdit TextEditor markup and all its wiring.
- ViewModel<->editor text sync (the manual two-way sync AvaloniaEdit needs
  because Text isn't a bindable AvaloniaProperty), plus feeding the live
  selection to the active tab for run-selection.
- Completion popup: ShowCompletion, ApplyFuzzyFilter, MaybeInsertTableAlias,
  the auto-open heuristics, and the session completion-recency.
- Syntax highlighting load + theme rewrite (ApplySqlHighlightingTheme),
  matching-bracket renderer, and current-line wash.
- Auto-close pairs, font-size zoom (Ctrl +/-/0, Ctrl+wheel), Shift+Enter
  run-current-statement, find/replace panel install + compact-template button
  wiring, and the drag-drop identifier drop target.
- Format-current-statement and expand-SELECT-* (the palette's Format/Expand
  actions and the Shift-F combos).

How it stays decoupled: the editor is genuinely window-central — it edits
whichever tab is active (MainViewModel.ActiveTab) and draws on
window-level services (the shared completion provider, the global word-wrap
preference, the Format/Expand-*/Find palette events), so the panel inherits
the host's MainViewModel as its DataContext rather than inventing a per-tab
sub-ViewModel that couldn't carry those. It tracks the active tab off its own
DataContext (decoupled from MainWindow's own active-tab tracking, which stays
for the results grid), and subscribes to the Format/Expand/Find events itself.
MainWindow keeps only what the window alone can drive: it resolves the panel
by name to hand off F6 focus and to open find/replace from OnKeyDown
(QueryEditor.FocusEditor / IsEditorFocused / OpenSearch). The cell inspector's
own JSON editor + its highlighting stay in MainWindow (they belong to the
results grid, extracted next).

No behavior change. Core stays Avalonia-free (all changes are in
PgNimbus.App). Build clean (0 warnings); all 519 Core tests pass.

Line counts: MainWindow.axaml.cs 2948 -> 2281, MainWindow.axaml 839 -> 833;
new QueryEditorPanel.axaml 22 + QueryEditorPanel.axaml.cs 835.
